完成处理日线数据的函数，准备测试。 准备将所有的索引变量的类型皆改为size_t。

+++ src/Data Container/ContainerFinnhubEconomicCalendar.h
@@ -22,6 +22,6 @@ public:
 
 protected:
 	vector<CEconomicCalendarPtr> m_vEconomicCalendar;
-	map<CString, long> m_mapEconomicCalendar;
-	long m_lLastTotalEconomicCalendar;
+	map<CString, size_t> m_mapEconomicCalendar;
+	size_t m_lLastTotalEconomicCalendar;
 };
